+++ package.json
@@ -9,21 +9,25 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "next": "^14.2.15",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "next": "14.2.16",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "sass": "^1.80.3",
     "uuid": "^10.0.0"
   },
   "devDependencies": {
     "@types/node": "^20.16.10",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/react": "18.3.12",
+    "@types/react-dom": "18.3.1",
     "@types/uuid": "^10.0.0",
     "eslint": "^8",
-    "eslint-config-next": "14.2.13",
+    "eslint-config-next": "14.2.16",
     "postcss": "^8",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
+  },
+  "overrides": {
+    "@types/react": "18.3.12",
+    "@types/react-dom": "18.3.1"
   }
 }
